Small little e2e test (#7)
* Durdling

* JFK its workin

* Refactoring to add complete

* Completes the request and prints result

Completes the request by calling the Complete method
on the Plugin and prints the returned pattern.
This allows to process the initial request and retrieve the completion.

* Refactors conformance plugin interaction

Moves the logic for handling acknowledge and complete requests
and responses into dedicated endpoint structures. This allows
for cleaner test setup and better control over the interaction
with the conformance plugin. The plugin now uses these endpoints
to send responses, improving the overall flow of conformance tests.

* Handle the error